pyproject.toml
[project]
name = "agent-adk-automated"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "google-adk>=1.0.0",
    "litellm>=1.71.1",
]
